+++ data_as_code/__init__.py
@@ -1,6 +1,7 @@
-from data_as_code._misc import ingredient
+from data_as_code import misc
 from data_as_code._recipe import Recipe
 from data_as_code._step import Step
+from data_as_code.misc import ingredient, source, intermediary, product
 from data_as_code.premade import source_local, source_http
 
 __version__ = '0.0.0'
