+++ tests/test_step.py
@@ -4,7 +4,7 @@ import pytest
 
 from data_as_code import exceptions as ex, ingredient
 from data_as_code._step import Step
-from data_as_code._misc import _Ingredient
+from data_as_code.misc import _Ingredient
 from data_as_code.premade import source_local
 from data_as_code._metadata import Metadata
